+++ mdfour.c
@@ -110,7 +110,7 @@ void mdfour_begin(struct mdfour *md)
 }
 
 
-static void mdfour_tail(const unsigned char *in, int n)
+static void mdfour_tail(const unsigned char *in, size_t n)
 {
 	unsigned char buf[128];
 	uint32_t M[16];
@@ -137,7 +137,7 @@ static void mdfour_tail(const unsigned char *in, int n)
 	}
 }
 
-void mdfour_update(struct mdfour *md, const unsigned char *in, int n)
+void mdfour_update(struct mdfour *md, const unsigned char *in, size_t n)
 {
 	uint32_t M[16];
 
@@ -149,7 +149,7 @@ void mdfour_update(struct mdfour *md, const unsigned char *in, int n)
 	}
 
 	if (md->tail_len) {
-		int len = 64 - md->tail_len;
+		size_t len = 64 - md->tail_len;
 		if (len > n) len = n;
 		memcpy(md->tail+md->tail_len, in, len);
 		md->tail_len += len;
